copying the color generator to stop purging
