+++ components/Posts/Posts.js
@@ -1,8 +1,18 @@
 import { Children } from "react";
 import Link from "next/link";
 
-import generateTagColors from "../../util/generateTagColors";
 export default function Posts({ posts }) {
+  const generateTagColors = (category) => {
+    const types = {
+      "WEB DEV": "pink",
+      BUSINESS: "blue",
+      UNDEFINED: "yellow",
+      GADGETS: "red",
+      NONSENSE: "indigo",
+      NEW: "green",
+    };
+    return types[category?.toUpperCase()] || types["UNDEFINED"];
+  };
   return (
     <div className="grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-3 gap-4 lg:gap-10">
       {Children.toArray(
